Workflow controls `on_exit` jumps to furthest module (#406)
This also reverts the starting module back to the Database module. This
will be important if entering Guided mode without an auto-loaded
database. In that case, the workflow should assume the starting module
*is* the Database module. It would be incorrect for the Data module to
be the final module.

Instead, on_exit in the workflow jumps to the furthest module ahead,
which should be the Data module, as the session will be unloaded and a
new one would need to be loaded to proceed from there.