WIP: Trying to nest non-post init calls
Materials, when constructed, successfully don't post_init run, but
caching the protocol (i.e. switching to another one) does call
post_init. Trying to figure out why that is. It shouldn't call post init
for Materials because we don't care if it's correct until trying to
officially create the protocol. Materials is an awkward situation
because it's part of an ABC Form creater, and then after choosing the
ABC, Materials are the value in a DictTableWidget. The DictTableWidget
has a special behavior where it can model complex value types, but the
post_init behavior isn't being properly suppressed when doing things
like caching, etc.