WIP: Prompting about saving session
Ran into a blocker that the slicer scene close takes over and closing
the scene will wipe out the session data, so closing the slicer scene
effectively makes the window impossible to show (as the session would be
wiped). Currently in search of ways to make this routine show before any
requests to close scene.

A separate issue will be disabling the scene save prompt altogether in
the custom app.