Throw if photo collection overwrite is declined (#445)
User interaction is prompted from the logic in the case that there is an
existing photocollection with the reference number requested to save.
This was likely implemented this way to mimic other implementations. It
is not the suggested approach, because the logic should not be tangled
with the GUI.

An exception is thrown if the user declines overwriting the
photocollection so that this information is propagated to other user
space functions.