Remove all photocollection ref number conflict warnings (#378)
Simply reset the list each time a session is loaded or a photocollection
is added.

Also check whether a photocollection reference number already in the
list before adding. The same procedure does not need to be done for
`OpenLIFUDataLogic.update_photocollections_affiliated_with_loaded_session`
because that sets the affiliated photocollections directly from the
updated database info.